control: properly use o_pause_iact to fix alternative dataflow
it's not used by the rs dataflow, but alternative rs needs it. thus it
has been incorrectly set to always-zero.
TODO: check if setting i_pause_iact=0 for address_generator is fine. If
yes, remove it from address_generator for both dataflows. If not,
introduce w_pause_iact and forward it accordingly.