control alt_rs: delay psum pass mode independent of i_enable
this introduces a small pipeline running independently of i_enable,
since the psum linebuffers do not stall and thus deliver output data
exactly 2 cycles after assertion of c_lb_read. otherwise, data may be
lost when throttling is enabled. maybe this is also required for the RS
dataflow when throttling is used.